+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 FaasJS use [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+- `@faasjs/utils`
+  - [Feature] Support literal and folded YAML block scalars in `parseYaml`.
+
 [`v8.0.0-beta.42 (2026-06-27)`](https://github.com/faasjs/faasjs/compare/v8.0.0-beta.41...v8.0.0-beta.42)
 
 - `@faasjs/dev`